+++ docs/branch-memory-bank/feature-add-test/activeContext.json
@@ -1,0 +1,80 @@
+{
+  "schema": "memory_document_v2",
+  "metadata": {
+    "id": "6d6232e5-d8df-40cc-a3cc-338f76473812",
+    "title": "アクティブコンテキスト",
+    "documentType": "active_context",
+    "path": "activeContext.json",
+    "tags": [
+      "active-context"
+    ],
+    "lastModified": "2025-04-07T01:48:07.680Z",
+    "createdAt": "2025-04-07T01:48:07.680Z",
+    "version": 1
+  },
+  "content": {
+    "currentWork": "現在取り組んでいる作業の説明",
+    "recentChanges": [
+      {
+        "date": "2025-04-07T01:48:07.680Z",
+        "description": "変更点1"
+      },
+      {
+        "date": "2025-04-07T01:48:07.680Z",
+        "description": "変更点2"
+      },
+      {
+        "date": "2025-04-07T01:48:07.680Z",
+        "description": "変更点3"
+      }
+    ],
+    "activeDecisions": [
+      {
+        "id": "11d2d283-1c08-49b8-a58b-f2d3290c1de0",
+        "description": "決定事項1"
+      },
+      {
+        "id": "853dd186-4cc5-42d3-a770-6a6ac95bfe50",
+        "description": "決定事項2"
+      },
+      {
+        "id": "2e943448-9a2a-4854-b8d8-498524329382",
+        "description": "決定事項3"
+      }
+    ],
+    "considerations": [
+      {
+        "id": "ace34867-e989-48a1-9912-8ddcadea9a76",
+        "description": "検討事項1",
+        "status": "open"
+      },
+      {
+        "id": "b5dad2b8-a98d-4ddf-ac83-1bd874f6ad18",
+        "description": "検討事項2",
+        "status": "open"
+      },
+      {
+        "id": "82ec496a-b87e-4b05-8740-e6af44eca1dc",
+        "description": "検討事項3",
+        "status": "open"
+      }
+    ],
+    "nextSteps": [
+      {
+        "id": "0a280f23-dcae-4aa2-9246-ee178b29d8f0",
+        "description": "次のステップ1",
+        "priority": "high"
+      },
+      {
+        "id": "1a54a9bb-1cdb-42c6-9bc8-5415e184ca20",
+        "description": "次のステップ2",
+        "priority": "medium"
+      },
+      {
+        "id": "ab23b2c1-d89c-47bc-8f20-8018d84a29a8",
+        "description": "次のステップ3",
+        "priority": "low"
+      }
+    ]
+  }
+}
