+++ docs/branch-memory-bank/feature-add-test/systemPatterns.json
@@ -1,0 +1,37 @@
+{
+  "schema": "memory_document_v2",
+  "metadata": {
+    "id": "5e81ea7b-09e9-437b-95a4-293c6f90d620",
+    "title": "システムパターン",
+    "documentType": "system_patterns",
+    "path": "systemPatterns.json",
+    "tags": [
+      "system-patterns"
+    ],
+    "lastModified": "2025-04-07T01:48:07.680Z",
+    "createdAt": "2025-04-07T01:48:07.680Z",
+    "version": 1
+  },
+  "content": {
+    "technicalDecisions": [
+      {
+        "id": "24ebcdb2-fc51-43e7-b868-9f2e701909c5",
+        "title": "決定事項のタイトル",
+        "context": "決定の背景や理由",
+        "decision": "具体的な決定内容",
+        "consequences": {
+          "positive": [
+            "影響1",
+            "影響2",
+            "影響3"
+          ],
+          "negative": []
+        },
+        "status": "proposed",
+        "date": "2025-04-07T01:48:07.680Z",
+        "alternatives": []
+      }
+    ],
+    "implementationPatterns": []
+  }
+}
