+++ docs/global-memory-bank/tags/index.json
@@ -9,8 +9,8 @@
       "index",
       "meta"
     ],
-    "lastModified": "2025-04-06T15:30:51.003Z",
-    "createdAt": "2025-04-06T15:30:51.003Z",
+    "lastModified": "2025-04-07T01:59:36.658Z",
+    "createdAt": "2025-04-07T01:59:36.658Z",
     "version": 1
   },
   "content": {
