+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ts-morph-mcp-server",
+  "name": "@t3ta/ts-introspect-mcp-server",
   "version": "0.1.0",
   "description": "A tool to introspect exported symbols from a TypeScript package using ts-morph and output a MCP-like schema.",
   "type": "module",
